// packages/sdk/src/typings.ts
export type PlatformStatus = string
// ...
export type BotFlags = string
// ...
export interface Bot {
    itag: string
    bot_id: string
    client_id: string
    extra_links: Link[]
    tags: string[]
    flags: BotFlags[]
    prefix: string
    user?: PlatformUser
    owner?: PlatformUser
    short: string
    long?: string
    library: string
    nsfw: boolean
    premium: boolean
    last_stats_post: string | null
    servers: number
    shards: number
    shard_list: number[]
    users: number
    votes: number
    clicks: number
    unique_clicks: number
    invite_clicks: number
    banner: string | null
    invite: string
    type: string
    vanity_ref: string
    vanity: string
    vote_banned: boolean
    start_premium_period: string | null
    premium_period_length: any
    cert_reason: string | null
    uptime: number
    total_uptime: number
    uptime_last_checked: string | null
    claimed_by: string | null
    approval_note: string | null
    created_at: string | null
    last_claimed: string | null
    legacy_webhooks: boolean
    team_owner?: Team
    captcha_opt_out: boolean
}
// ...
export interface Link {
    name: string
    value: string
}
// ...
export interface PlatformUser {
    id: string
    username: string
    display_name: string
    avatar: string
    bot: boolean
    status: PlatformStatus
    extra_data: { [key: string]: any }
}
// ...
export interface Team {
    id: string
    name: string
    avatar: string
    banner: string | null
    short: string | null
    tags: string[]
    votes: number
    extra_links: Link[]
    entities?: TeamEntities
}
export interface TeamBulkFetch {
    teams: Team[]
}
export interface TeamEntities {
    targets?: string[]
    members?: TeamMember[]
    bots?: IndexBot[]
}